requirements.txt
# torch ecosystem
torch
torchvision 
torchaudio 

# stable diffusion
diffusers
transformers
accelerate
peft
huggingface-hub
safetensors

# misc
numpy
opencv-python
pillow
tqdm
pyyaml
flask
